Renamed `index.tsx` files for clarity
