+++ components/ComicViewer/NavigationBtns/NavigationBtns.tsx
